requirements.txt
amqp==2.4.2
billiard==3.5.0.5
celery==4.1.1
Django==1.11.20
django-appconf==1.0.3
django-crispy-forms==1.7.2
django-filter==2.1.0
django-imagekit==4.0.2
django-search-listview==0.2.1
django-widget-tweaks==1.4.3
Faker==1.0.4
html5lib==1.0.1
kombu==4.5.0
pilkit==2.0
Pillow==5.4.1
psycopg2-binary==2.7.7
PyPDF2==1.26.0
python-dateutil==2.8.0
pytz==2018.9
redis==3.2.1
reportlab==3.5.17
six==1.12.0
text-unidecode==1.2
vine==1.3.0
webencodings==0.5.1
xhtml2pdf==0.2.3
